move details form to sub directory
